move configuration into a single class
